Allow all candidate instructions to be examined
Previously only the candidate instructions for a destination unit were
shortlisted to the number of available spaces in the destination unit.
This caused some eligible candidate instructions to be skipped if some
of the shortlisted candidates didn't make it to the destination unit.

I removed the shortlisting, effectively making all candidates available
for picking by the destination unit.