Decouple test inputs from expected results
Parameters of parameterized test cases were found to derive both the
values for the test inputs as well as the test expected results. This
opens the door for providing parallel implementations of the
functionality under the test. Test inputs and corresponding expected
results should be provided as immediate values, without derivation
logic.