Install coveralls dependencies
Coveralls GitHub Action depends on the latest published release of the
Coveralls Universal Coverage Reporter. Up to release 0.6.9 of the
Coveralls Universal Coverage Reporter, the reporter was able to process
.coverage files without relying on the python coverage executable being
present. Come release 0.6.10 and now the reporter requires the python
coverage executable to be present. That's why I'm installing the python
coverage package before running the coveralls action.